requirements.txt
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
Django==2.1.1
django-filter==2.0.0
django-fsm==2.6.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.7
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
psycopg2==2.7.5
pytz==2018.5
requests==2.19.1
ruamel.yaml==0.15.71
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
